+++ GDEdit/GDE.App/Main/GDEApp.cs
@@ -1,4 +1,4 @@
-﻿using GDE.App.Main.Containers;
+﻿using GDE.App.Main.Graphics.Containers;
 using GDE.App.Main.Screens.Menu;
 using GDE.App.Main.Toasts;
 using GDE.App.Main.Tools;
