+++ GDEdit/GDE.App/Main/Graphics/User Interface/LCDComponents/LCDNumber.cs
@@ -3,7 +3,7 @@ using osu.Framework.Graphics.Containers;
 using osuTK;
 using System;
 
-namespace GDE.App.Main.UI.LCDComponents
+namespace GDE.App.Main.Graphics.UserInterface.LCDComponents
 {
     public class LCDNumber : FillFlowContainer
     {
