+++ GDEdit/GDE.App/Main/Graphics/User Interface/Toolbar/ToolbarMenu.cs
@@ -1,7 +1,6 @@
-﻿using osu.Framework.Allocation;
+﻿using GDE.App.Main.Colors;
+using osu.Framework.Allocation;
 using osu.Framework.Audio;
-using osu.Framework.Audio.Sample;
-using osuTK.Graphics;
 using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
@@ -9,13 +8,13 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input.Events;
 using osuTK;
-using GDE.App.Main.Colors;
+using osuTK.Graphics;
 
-namespace GDE.App.Main.UI
+namespace GDE.App.Main.Graphics.UserInterface.Toolbar
 {
-    public class GDEMenu : Menu
+    public class ToolbarMenu : Menu
     {
-        public GDEMenu(Direction direction, bool topLevelMenu = false)
+        public ToolbarMenu(Direction direction, bool topLevelMenu = false)
             : base(direction, topLevelMenu)
         {
             BackgroundColour = Color4.Black.Opacity(0.5f);
@@ -41,22 +40,29 @@ namespace GDE.App.Main.UI
             }
         }
 
-        protected override Menu CreateSubMenu() => new GDEMenu(Direction.Vertical)
+        protected override DrawableMenuItem CreateDrawableMenuItem(MenuItem item) => new DrawableToolbarMenu(item);
+
+        protected override Menu CreateSubMenu() => new ToolbarMenu(Direction.Vertical)
         {
             Anchor = Direction == Direction.Horizontal ? Anchor.BottomLeft : Anchor.TopRight
         };
 
-        protected class DrawableGDEMenuItem : DrawableMenuItem
+        protected class DrawableToolbarMenu : DrawableMenuItem
         {
-            private const int margin_horizontal = 10;
+            private const int margin_horizontal = 17;
             private const int text_size = 17;
             private const int transition_length = 80;
             public const int MARGIN_VERTICAL = 4;
 
             private TextContainer text;
 
-            public DrawableGDEMenuItem(MenuItem item)
+            public DrawableToolbarMenu(MenuItem item)
                 : base(item)
+            {
+            }
+
+            [BackgroundDependencyLoader]
+            private void load(AudioManager audio)
             {
                 BackgroundColour = Color4.Transparent;
                 BackgroundColourHover = GDEColors.FromHex(@"172023");
@@ -66,19 +72,17 @@ namespace GDE.App.Main.UI
 
             private void updateTextColour()
             {
-                switch ((Item as GDEMenuItem)?.Type)
+                switch ((Item as ToolbarMenuItem)?.Type)
                 {
-                    default:
-                    case MenuItemType.Standard:
-                        text.Colour = Color4.White;
-                        break;
-
                     case MenuItemType.Destructive:
                         text.Colour = Color4.Red;
                         break;
-
                     case MenuItemType.Highlighted:
                         text.Colour = GDEColors.FromHex(@"ffcc22");
+                        break;
+                    case MenuItemType.Standard:
+                    default:
+                        text.Colour = Color4.White;
                         break;
                 }
             }
@@ -137,7 +141,7 @@ namespace GDE.App.Main.UI
                             Alpha = 0,
                             Anchor = Anchor.CentreLeft,
                             Origin = Anchor.CentreLeft,
-                            Font = new FontUsage(size: text_size, weight: "Bold"),
+                            Font = new FontUsage(size: text_size),
                             Margin = new MarginPadding { Horizontal = margin_horizontal, Vertical = MARGIN_VERTICAL },
                         }
                     };
