+++ GDEdit/GDE.App/Main/Levels/LevelPreview.cs
@@ -1,8 +1,8 @@
-﻿using GDE.App.Main.Containers;
+﻿using GDE.App.Main.Graphics.Containers;
 using GDE.App.Main.Objects;
 using GDE.App.Main.Screens.Edit;
 using GDE.App.Main.Screens.Edit.Components;
-using GDE.App.Main.UI;
+using GDE.App.Main.Graphics.UserInterface;
 using GDEdit.Application;
 using GDEdit.Application.Editor;
 using GDEdit.Utilities.Objects.GeometryDash;
