+++ GDEdit/GDE.App/Main/Screens/Edit/Components/ObjectAdditionPanel.cs
@@ -1,33 +1,12 @@
 ﻿using GDE.App.Main.Colors;
-using GDE.App.Main.Containers;
 using GDE.App.Main.Objects;
-using GDE.App.Main.Overlays;
 using GDE.App.Main.Panels;
-using GDE.App.Main.Screens.Menu.Components;
-using GDE.App.Main.Tools;
-using GDE.App.Main.UI;
-using GDEdit.Application;
-using GDEdit.Application.Editor;
-using GDEdit.Utilities.Objects.GeometryDash;
 using GDEdit.Utilities.Objects.GeometryDash.LevelObjects;
-using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
-using osu.Framework.Graphics.Colour;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.Transforms;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Bindings;
-using osu.Framework.Input.Events;
-using osu.Framework.Logging;
-using osu.Framework.Screens;
 using osuTK;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Threading.Tasks;
 
 namespace GDE.App.Main.Screens.Edit.Components
 {
