+++ GDEdit/GDE.App/Main/Screens/Menu/Components/LevelList.cs
@@ -1,22 +1,13 @@
 ﻿using GDE.App.Main.Colors;
-using GDE.App.Main.Containers;
-using GDE.App.Main.Overlays;
-using GDE.App.Main.Screens.Menu.Components;
-using GDE.App.Main.Tools;
-using GDE.App.Main.UI;
 using GDEdit.Application;
 using GDEdit.Utilities.Functions.Extensions;
 using GDEdit.Utilities.Objects.GeometryDash;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Bindings;
 using osu.Framework.Logging;
-using osu.Framework.Screens;
 using osuTK;
 using System;
 using System.Collections.Generic;
